WIP upgrade runbot to Odoo 11.0
