--- runbot/__openerp__.py
@@ -1,20 +1,0 @@
-{
-    'name': 'Runbot',
-    'category': 'Website',
-    'summary': 'Runbot',
-    'version': '1.3',
-    'description': "Runbot",
-    'author': 'Odoo SA',
-    'depends': ['website', 'base_setup'],
-    'external_dependencies': {
-        'python': ['matplotlib'],
-    },
-    'data': [
-        'runbot.xml',
-        'res_config_view.xml',
-        'security/runbot_security.xml',
-        'security/ir.model.access.csv',
-        'security/ir.rule.csv',
-    ],
-    'installable': True,
-}
