+++ runbot/data/runbot_cron.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding='UTF-8'?>
+<odoo>
+	<record model="ir.cron" id="runbot_repo_cron">
+        <field name="name">Runbot Cron</field>
+        <field name="model_id" ref="model_runbot_repo"/>
+        <field name="state">code</field>
+        <field name="code">model._cron()</field>
+        <field name="interval_number">1</field>
+        <field name="interval_type">minutes</field>
+        <field name="numbercall">-1</field>
+        <field name="doall" eval="False"/>
+    </record>
+</odoo>
