+++ runbot/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from . import repo, branch, build, event
+from . import res_config_settings
