--- runbot/res_config_view.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="view_runbot_configuration" model="ir.ui.view">
-            <field name="name">Configure Runbot</field>
-            <field name="model">runbot.config.settings</field>
-            <field name="arch" type="xml">
-                <form string="Configure Runbot" class= "oe_form_configuration" version="7.0">
-                    <header>
-                        <button string="Apply" type="object" name="execute" class="oe_highlight"/>
-                        or
-                        <button string="Cancel" type="object" name="cancel" class="oe_link"/>
-                    </header>
-                    <separator string="Runbot"/>
-                    <group>
-                        <label for="id" string="Settings"/>
-                        <div>
-                            <div>
-                                <field name="default_workers" class="oe_inline"/>
-                                <label for="default_workers"/>
-                            </div>
-                            <div>
-                                <field name="default_running_max" class="oe_inline"/>
-                                <label for="default_running_max"/>
-                            </div>
-                            <div>
-                                <field name="default_timeout" class="oe_inline"/>
-                                <label for="default_timeout"/>
-                            </div>
-                            <div>
-                                <field name="default_starting_port" class="oe_inline"/>
-                                <label for="default_starting_port"/>
-                            </div>
-                            <div>
-                                <field name="default_domain" class="oe_inline"/>
-                                <label for="default_domain"/>
-                            </div>
-                        </div>
-                    </group>
-                </form>
-            </field>
-        </record>
-        <record id="action_runbot_configuration" model="ir.actions.act_window">
-            <field name="name">Configure Runbot</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">runbot.config.settings</field>
-            <field name="view_mode">form</field>
-            <field name="target">inline</field>
-        </record>
-        <menuitem id="menu_runbot_configuration" name="Runbot" parent="base.menu_config"
-            sequence="19" action="action_runbot_configuration"/>
-
-    </data>
-</openerp>
