+++ runbot/views/res_config_settings_views.xml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <record id="res_config_settings_view_form" model="ir.ui.view">
+            <field name="name">res.config.settings.view.form.inherit.runbot</field>
+            <field name="model">res.config.settings</field>
+            <field name="inherit_id" ref="base.res_config_settings_view_form"/>
+            <field name="arch" type="xml">
+              <xpath expr="//div[hasclass('settings')]" position="inside">
+                <div class="app_settings_block" data-string="Runbot" string="Runbot" data-key="runbot">
+                  <h2>Runbot configuration</h2>
+                  <div class="row mt16 o_settings_container">
+                    <div class="col-xs-12 col-md-6 o_setting_box">
+                      <div class="o_setting_right_pane">
+                        <label for="runbot_root"/>
+                        <field name="runbot_root"/>
+                        <label for="runbot_workers"/>
+                        <field name="runbot_workers"/>
+                        <label for="runbot_running_max"/>
+                        <field name="runbot_running_max"/>
+                      </div>
+                       <div class="o_setting_right_pane">
+                        <label for="runbot_timeout"/>
+                        <field name="runbot_timeout"/>
+                        <label for="runbot_starting_port"/>
+                        <field name="runbot_starting_port"/>
+                        <label for="runbot_domain"/>
+                        <field name="runbot_domain"/>
+                        <label for="runbot_max_age"/>
+                        <field name="runbot_max_age"/>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </xpath>
+            </field>
+        </record>
+
+        <record id="action_runbot_configuration" model="ir.actions.act_window">
+            <field name="name">Settings</field>
+            <field name="type">ir.actions.act_window</field>
+            <field name="res_model">res.config.settings</field>
+            <field name="view_mode">form</field>
+            <field name="target">inline</field>
+            <field name="context">{'module' : 'runbot'}</field>
+        </record>
+
+        <menuitem id="menu_runbot_global_settings" name="Settings"
+            parent="runbot_menu_root" sequence="0" action="action_runbot_configuration" groups="base.group_system"/>
+    </data>
+</odoo>
